A little bit refactor
