+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import App from "./components/app/app.jsx";
 import configureAPI from "./api.js";
 
-import finalReducer from "./reducer/index.js";
+import reducer from "./reducer/reducer.js";
 import {Operation} from "./reducer/data/data";
 import {Provider} from "react-redux";
 import {createStore, applyMiddleware} from "redux";
@@ -14,7 +14,7 @@ const init = () => {
   const api = configureAPI((...args) => store.dispatch(...args));
 
   const store = createStore(
-      finalReducer,
+      reducer,
       compose(
           applyMiddleware(thunk.withExtraArgument(api)),
           window.__REDUX_DEVTOOLS_EXTENSION__ ? window.__REDUX_DEVTOOLS_EXTENSION__() : (f) => f
