+++ src/reducer/data/data.js
@@ -69,7 +69,7 @@ const reducer = (state = initialState, action) => {
       });
   }
 
-  return Object.assign({}, state);
+  return state;
 };
 
 export {ActionType, ActionCreator, Operation, reducer};
