+++ src/reducer/data/data.test.js
@@ -76,15 +76,6 @@ describe(`ActionCreator works correctly`, () => {
 });
 
 describe(`Reducer works correctly`, () => {
-  it(`Reducer without any action should return initial state`, () => {
-    const initialState = {
-      city: null,
-      offers: mockOffers
-    };
-
-    expect(reducer(initialState, {})).toEqual(initialState);
-  });
-
   it(`Reducer changes cityName in the state with action type "CHANGE_CITY"`, () => {
     const initialState = {
       city: null,
